Propagate samplingProbability from persisted events to backend senders
Use event.samplingProbability instead of hardcoded value when building API
and pixel payloads. Change GlobalSection.sampleRate type from Int
to Float to match the stored value.

+++ statistics/statistics-impl/src/main/java/com/duckduckgo/app/statistics/wideevents/api/WideEventRequest.kt
@@ -28,7 +28,7 @@ data class WideEventRequest(
 data class GlobalSection(
     @field:Json(name = "platform") val platform: String,
     @field:Json(name = "type") val type: String,
-    @field:Json(name = "sample_rate") val sampleRate: Int,
+    @field:Json(name = "sample_rate") val sampleRate: Float,
 )
 
 data class AppSection(
